test: fix checkstlye
