+++ src/test/java/com/xenoterracide/gradle/semver/SemverExtensionTest.java
@@ -4,10 +4,6 @@
 package com.xenoterracide.gradle.semver;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.mockito.Mockito.never;
-import static org.mockito.Mockito.spy;
-import static org.mockito.Mockito.times;
-import static org.mockito.Mockito.verify;
 
 import java.io.File;
 import java.util.Collections;
@@ -71,9 +67,7 @@ class SemverExtensionTest {
         .isGreaterThan(v010BldV2)
         .extracting(Semver::getVersion, Semver::toString)
         .allSatisfy(o -> {
-          assertThat(o)
-            .asInstanceOf(InstanceOfAssertFactories.STRING)
-            .startsWith("0.1.1-alpha.2+2.g");
+          assertThat(o).asInstanceOf(InstanceOfAssertFactories.STRING).startsWith("0.1.1-alpha" + ".2+2.g");
         });
 
       git.tag().setName("v0.1.1-rc.1").call();
@@ -99,13 +93,7 @@ class SemverExtensionTest {
         .isGreaterThan(v010)
         .isGreaterThan(v000)
         .hasToString("0.1.1")
-        .extracting(
-          Semver::getMajor,
-          Semver::getMinor,
-          Semver::getPatch,
-          Semver::getPreRelease,
-          Semver::getBuild
-        )
+        .extracting(Semver::getMajor, Semver::getMinor, Semver::getPatch, Semver::getPreRelease, Semver::getBuild)
         .containsExactly(0, 1, 1, Collections.emptyList(), Collections.emptyList());
 
       assertThat(VersionNumber.parse(v010BldV2.toString()))
@@ -176,13 +164,7 @@ class SemverExtensionTest {
         .isGreaterThan(v010)
         .isGreaterThan(v000)
         .hasToString("0.1.1")
-        .extracting(
-          Semver::getMajor,
-          Semver::getMinor,
-          Semver::getPatch,
-          Semver::getPreRelease,
-          Semver::getBuild
-        )
+        .extracting(Semver::getMajor, Semver::getMinor, Semver::getPatch, Semver::getPreRelease, Semver::getBuild)
         .containsExactly(0, 1, 1, Collections.emptyList(), Collections.emptyList());
 
       assertThat(VersionNumber.parse(v010BldV2.toString()))
@@ -197,103 +179,6 @@ class SemverExtensionTest {
         .isEqualByComparingTo(new ComparableVersion(v010BldV3.toString()))
         .isLessThan(new ComparableVersion(v011.toString()));
     }
-  }
-
-  @Test
-  void getMavenAlpha() throws Exception {
-    try (var git = Git.init().setDirectory(projectDir).call()) {
-      git.commit().setMessage("initial commit").call();
-      var pg = new SemverExtension(() -> Optional.of(git));
-
-      var v000 = pg.getMavenAlpha();
-      assertThat(v000).extracting(Semver::getVersion).isEqualTo("0.0.1-alpha.1000000000000000");
-      assertThat(v000)
-        .hasToString("0.0.1-alpha.1000000000000000")
-        .extracting(Semver::getMajor, Semver::getMinor, Semver::getPatch, Semver::getPreRelease)
-        .containsExactly(0, 0, 1, List.of("alpha", "1000000000000000"));
-
-      git.tag().setName("v0.1.0").call();
-      var v010 = pg.getMavenAlpha();
-      git.commit().setMessage("second commit").call();
-
-      var v010BldV2 = pg.getMavenAlpha();
-
-      assertThat(v010BldV2)
-        .extracting(Semver::getVersion, Semver::toString)
-        .allSatisfy(o -> {
-          assertThat(o)
-            .asInstanceOf(InstanceOfAssertFactories.STRING)
-            .startsWith("0.1.1-alpha.1001")
-            .matches("^0\\.1\\.1-alpha.1001\\p{XDigit}{12}$");
-        });
-
-      git.commit().setMessage("third commit").call();
-
-      var v010BldV3 = pg.getMavenAlpha();
-
-      assertThat(v010BldV3)
-        .isGreaterThan(v000)
-        .isGreaterThan(v010)
-        .isGreaterThan(v010BldV2)
-        .extracting(Semver::getVersion, Semver::toString)
-        .allSatisfy(o -> {
-          assertThat(o)
-            .asInstanceOf(InstanceOfAssertFactories.STRING)
-            .startsWith("0.1.1-alpha.1002");
-        });
-
-      git.tag().setName("v0.1.1-rc.1").call();
-
-      var v011Rc1 = pg.getMavenAlpha();
-
-      assertThat(v011Rc1)
-        .isGreaterThan(v000)
-        .isGreaterThan(v010)
-        .isGreaterThan(v010BldV2)
-        .isGreaterThan(v010BldV3)
-        .extracting(Semver::getVersion, Semver::toString)
-        .allSatisfy(o -> {
-          assertThat(o).asInstanceOf(InstanceOfAssertFactories.STRING).startsWith("0.1.1-rc.1");
-        });
-
-      git.tag().setName("v0.1.1").call();
-
-      var v011 = pg.getMavenAlpha();
-
-      assertThat(v011)
-        .isGreaterThan(v010BldV2)
-        .isGreaterThan(v010)
-        .isGreaterThan(v000)
-        .hasToString("0.1.1")
-        .extracting(
-          Semver::getMajor,
-          Semver::getMinor,
-          Semver::getPatch,
-          Semver::getPreRelease,
-          Semver::getBuild
-        )
-        .containsExactly(0, 1, 1, Collections.emptyList(), Collections.emptyList());
-
-      assertThat(VersionNumber.parse(v010BldV2.toString()))
-        .isGreaterThan(VersionNumber.parse(v000.toString()))
-        .isGreaterThan(VersionNumber.parse(v010.toString()))
-        .isLessThan(VersionNumber.parse(v010BldV3.toString()))
-        .isLessThan(VersionNumber.parse(v011.toString()));
-
-      assertThat(new ComparableVersion(v010BldV2.toString()))
-        .isGreaterThan(new ComparableVersion(v000.toString()))
-        .isGreaterThan(new ComparableVersion(v010.toString()))
-        .isLessThan(new ComparableVersion(v010BldV3.toString()))
-        .isLessThan(new ComparableVersion(v011.toString()));
-    }
-  }
-
-  @Test
-  void mavenAlgorithm() {
-    var pg = spy(new SemverExtension(() -> Optional.empty()));
-    assertThat(pg.getMaven()).hasToString("0.0.0-SNAPSHOT");
-    verify(pg, times(1)).getMavenSnapshot();
-    verify(pg, never()).getMavenAlpha();
   }
 
   @Test
